Use subtests inside of TestSumOfPrimesInRange
This test didn't inlclude a for loop to iterate through the various test
cases. As a result it only ran one test, which checked the sum of primes
in the range of 10 to 19.

After correctly running all tests it was discovered that function had a
bug which always returned a sum of 60 because the call to PrimesInRange
received hardcoded values instead of the start and end function parameters.

Closes: #1
See also: #5