Remove hardcoded values from call to PrimesInRange
A call to PrimesInRange used hardcoded values. This resulted in the
function always iterating over the same slice of prime numbers and
always returing 60 as the sum of primes for a given range.

This bug was missed because when the test was originally written, it
ommitted iterating over the various test cases.

Closes: #5
See also: #1