Add error return value to range functions
Previously these functions simply returned a slice or an int value. The
introduction of the StartGreaterThanEndError allows for the start and end
input parameters to be checked. If the start parameter is greater than
or equal to the end paramater, then these functions should return either
a nil slice or -1, depending on the return type and the
StartGreaterThanEndError to signal that a wrong input has been passed to
the function.